* Fixed all Android GFX bugs, fixed light_base.py

+++ src/resources/base_android.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ \! -d base ] ; then
+	echo "You should run this command from the UfoAI root dir, like: src/resources/base_android.sh"
+	exit 1
+fi
+
 rm -rf base_light base_full 0base.pk3 0music.pk3 1base.pk3
 src/resources/light_base.py
 
@@ -15,8 +20,6 @@ mv -f downsampledimages-1.txt base_full/downsampledimages.txt
 cd base_light
 cp -f ../base/android.cfg config.cfg
 cp -f ufos/android/* ufos
-zip -r -n .png:.jpg:.ogg:.ogm ../0music.pk3 sound
-rm -rf sound
 zip -r -n .png:.jpg:.ogg:.ogm ../0base.pk3 *
 cd ../base_full
 zip -r -n .png:.jpg:.ogg:.ogm ../0music.pk3 music
